--- Diary/Controller/DiaryDetailViewController.swift
@@ -1,14 +1,0 @@
-//
-//  DiaryDetailViewController.swift
-//  Diary
-//
-//  Created by 백곰, 주디 on 2022/08/17.
-//
-
-import UIKit
-
-final class DiaryDetailViewController: UIViewController {
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
