--- Diary/View/DiaryDetailView.swift
@@ -1,12 +1,0 @@
-//
-//  DiaryDetailView.swift
-//  Diary
-//
-//  Created by 백곰,주디 on 2022/08/17.
-//
-
-import UIKit
-
-final class DiaryDetailView: UIView {
-    
-}
